+++ gamse/pipelines/xinglong216hrs/__init__.py
@@ -1,0 +1,1485 @@
+import os
+import re
+import shutil
+import datetime
+import logging
+logger = logging.getLogger(__name__)
+import dateutil.parser
+
+import numpy as np
+from scipy.signal import savgol_filter
+from scipy.ndimage.filters import gaussian_filter
+from scipy.interpolate import InterpolatedUnivariateSpline
+import astropy.io.fits as fits
+from astropy.table import Table
+from astropy.time import Time
+import matplotlib.pyplot as plt
+import matplotlib.ticker as tck
+import matplotlib.dates as mdates
+
+from ..echelle.imageproc import (combine_images, array_to_table,
+                                 fix_pixels)
+from ..echelle.trace import find_apertures, load_aperture_set, TraceFigureCommon
+from ..echelle.flat  import get_fiber_flat, mosaic_flat_auto, mosaic_images
+from ..echelle.extract import extract_aperset
+from ..echelle.wlcalib import (wlcalib, recalib, get_calib_from_header,
+                               get_time_weight, find_caliblamp_offset,
+                               reference_wavelength,
+                               reference_self_wavelength,
+                               #self_reference_singlefiber,
+                               #wl_reference,
+                               )
+from ..echelle.background import find_background
+from ..utils.onedarray import get_local_minima
+from ..utils.regression import iterative_polyfit
+from ..utils.obslog import parse_num_seq, read_obslog
+from ..utils.misc import extract_date
+from .common import plot_background_aspect1, FormattedInfo
+from .reduction          import Reduction
+
+def get_badpixel_mask(shape, bins):
+    """Get the mask of bad pixels and columns.
+
+    Args:
+        shape (tuple): Shape of image.
+        bins (tuple): CCD bins.
+
+    Returns:
+        :class:`numpy.ndarray`: 2D binary mask, where bad pixels are marked with
+            *True*, others *False*.
+
+    The bad pixels are found *empirically*.
+        
+    """
+    mask = np.zeros(shape, dtype=np.bool)
+    if bins == (1, 1) and shape == (4136, 4096):
+        h, w = shape
+
+        mask[349:352, 627:630] = True
+        mask[349:h//2, 628]    = True
+
+        mask[1604:h//2, 2452] = True
+
+        mask[280:284,3701]   = True
+        mask[274:h//2, 3702] = True
+        mask[272:h//2, 3703] = True
+        mask[274:282, 3704]  = True
+
+        mask[1720:1722, 3532:3535] = True
+        mask[1720, 3535]           = True
+        mask[1722, 3532]           = True
+        mask[1720:h//2,3533]       = True
+
+        mask[347:349, 4082:4084] = True
+        mask[347:h//2,4083]      = True
+
+        mask[h//2:2631, 1909] = True
+    else:
+        print('No bad pixel information for this CCD size.')
+        raise ValueError
+    return mask
+
+def get_mask(data, head):
+    """Get the mask of input image.
+
+    Args:
+        data (:class:`numpy.ndarray`): Input image data.
+        head (:class:`astropy.io.fits.Header`): Input FITS header.
+
+    Returns:
+        :class:`numpy.ndarray`: Image mask.
+
+    The shape of output mask is determined by the keywords in the input FITS
+    header. The numbers of columns and rows are given by::
+     
+        N (columns) = head['NAXIS1'] - head['COVER']
+
+        N (rows)    = head['NAXIS2'] - head['ROVER']
+
+    where *head* is the input FITS header. 
+
+    """
+
+    saturation_adu = 65535
+
+    # determine shape of output image (also the shape of science region)
+    y1 = head.get('CRVAL2', 0)
+    rover = head.get('ROVER', 0)
+    y2 = y1 + head['NAXIS2'] - rover
+    x1 = head.get('CRVAL1', 0)
+    cover = head.get('COVER', 64)
+    x2 = x1 + head['NAXIS1'] - cover
+    newshape = (y2-y1, x2-x1)
+
+    # find the saturation mask
+    mask_sat = (data[y1:y2, x1:x2] >= saturation_adu)
+    # get bad pixel mask
+    rbin = head.get('RBIN', 1)
+    cbin = head.get('CBIN', 1)
+    bins = (rbin, cbin)
+    mask_bad = get_badpixel_mask(newshape, bins=bins)
+
+    mask = np.int16(mask_sat)*4 + np.int16(mask_bad)*2
+
+    return mask
+
+
+def correct_overscan(data, head, mask=None):
+    """Correct overscan for an input image and update related information in the
+    FITS header.
+    
+    Args:
+        data (:class:`numpy.ndarray`): Input image data.
+        head (:class:`astropy.io.fits.Header`): Input FITS header.
+        mask (:class:`numpy.ndarray`): Input image mask.
+    
+    Returns:
+        tuple: A tuple containing:
+
+            * **data** (:class:`numpy.ndarray`) – Output image with overscan
+              corrected.
+            * **card_lst** (*list*) – A new card list for FITS header.
+            * **overmean** (*float) – Mean value of overscan pixels.
+    """
+    # define the cosmic ray fixing function
+    def fix_cr(data):
+        m = data.mean(dtype=np.float64)
+        s = data.std(dtype=np.float64)
+        _mask = data > m + 3.*s
+        if _mask.sum()>0:
+            x = np.arange(data.size)
+            f = InterpolatedUnivariateSpline(x[~_mask], data[~_mask], k=3)
+            return f(x)
+        else:
+            return data
+
+    h, w = data.shape
+    cover = head.get('COVER', 64)
+    x1, x2 = w-cover, w
+
+    # find the overscan level along the y-axis
+    ovr_lst1 = data[0:h//2,x1+2:x2].mean(dtype=np.float64, axis=1)
+    ovr_lst2 = data[h//2:h,x1+2:x2].mean(dtype=np.float64, axis=1)
+
+    ovr_lst1_fix = fix_cr(ovr_lst1)
+    ovr_lst2_fix = fix_cr(ovr_lst2)
+
+    # apply the sav-gol fitler to the mean of overscan
+    winlength = 301
+    polyorder = 3
+    ovrsmooth1 = savgol_filter(ovr_lst1_fix,
+                    window_length=winlength, polyorder=polyorder)
+    ovrsmooth2 = savgol_filter(ovr_lst2_fix,
+                    window_length=winlength, polyorder=polyorder)
+
+    # determine shape of output image (also the shape of science region)
+    y1 = head.get('CRVAL2', 0)
+    rover = head.get('ROVER', 0)
+    y2 = y1 + head['NAXIS2'] - rover
+    ymid = (y1 + y2)//2
+    x1 = head.get('CRVAL1', 0)
+    cover = head.get('COVER', 64)
+    x2 = x1 + head['NAXIS1'] - cover
+    newshape = (y2-y1, x2-x1)
+
+    # subtract overscan
+    new_data = np.zeros(newshape, dtype=np.float64)
+    ovrdata1 = np.repeat([ovrsmooth1],x2-x1,axis=0).T
+    ovrdata2 = np.repeat([ovrsmooth2],x2-x1,axis=0).T
+    new_data[y1:ymid, x1:x2] = data[y1:ymid,x1:x2] - ovrdata1
+    new_data[ymid:y2, x1:x2] = data[ymid:y2,x1:x2] - ovrdata2
+    overmean = (ovrsmooth1.mean() + ovrsmooth2.mean())/2.
+
+    if mask is not None:
+        # fix bad pixels
+        bad_mask = (mask&2 > 0)
+        new_data = fix_pixels(new_data, bad_mask, 'x', 'linear')
+
+    card_lst = []
+    prefix = 'HIERARCH GAMSE OVERSCAN '
+    card_lst.append((prefix + 'CORRECTED', True))
+    card_lst.append((prefix + 'METHOD',    'smooth:savgol'))
+    card_lst.append((prefix + 'WINLEN',    winlength))
+    card_lst.append((prefix + 'POLYORDER', polyorder))
+    #card_lst.append((prefix+' AXIS-1',    '{}:{}'.format(x1, x2)))
+    #card_lst.append((prefix+' AXIS-2',    '{}:{}'.format()))
+
+    return new_data, card_lst, overmean
+
+
+def parse_bias_frames(logtable, config, pinfo):
+    """Parse the bias images and return the bias as an array.
+
+    Args:
+        logtable ():
+        config ():
+        pinfo ():
+
+    Returns:
+        bias:
+        bias_card_lst:
+
+    """
+
+    rawdata = config['data']['rawdata']
+    section = config['reduce.bias']
+    bias_file = section['bias_file']
+
+    bias_data_lst = []
+    bias_fileid_lst = []
+
+    for logitem in logtable:
+        if logitem['object'].strip().lower()=='bias':
+            filename = os.path.join(rawdata, logitem['fileid']+'.fits')
+            data, head = fits.getdata(filename, header=True)
+            mask = get_mask(data, head)
+            data, card_lst, overmean = correct_overscan(data, head, mask)
+
+            # print info
+            if len(bias_data_lst) == 0:
+                print('* Combine Bias Images: {}'.format(bias_file))
+                print(' '*2 + pinfo.get_title())
+                print(' '*2 + pinfo.get_separator())
+            print(' '*2 + pinfo.get_format().format(logitem, overmean))
+
+            bias_data_lst.append(data)
+            bias_fileid_lst.append(logitem['fileid'])
+
+    # get number of bias images
+    n_bias = len(bias_data_lst)
+
+    prefix = 'HIERARCH GAMSE BIAS '
+    bias_card_lst = []
+    bias_card_lst.append((prefix + 'NFILE', n_bias))
+
+    if n_bias == 0:
+        # there is no bias frames
+        bias = None
+    else:
+        for i, fileid in enumerate(bias_fileid_lst):
+            bias_card_lst.append((prefix + 'FILEID {:02d}'.format(i), fileid))
+
+        # there is bias frames
+        print(' '*2 + pinfo.get_separator())
+
+        # combine bias images
+        bias_data_lst = np.array(bias_data_lst)
+
+        bias = combine_images(bias_data_lst,
+                mode       = 'mean',
+                upper_clip = section.getfloat('cosmic_clip'),
+                maxiter    = section.getint('maxiter'),
+                mask       = (None, 'max')[n_bias>=3],
+                )
+
+        ############## bias smooth ##################
+        if section.getboolean('smooth'):
+            # bias needs to be smoothed
+            smooth_method = section.get('smooth_method')
+
+            h, w = bias.shape
+            if smooth_method in ['gauss','gaussian']:
+                # perform 2D gaussian smoothing
+                smooth_sigma = section.getint('smooth_sigma')
+                smooth_mode  = section.get('smooth_mode')
+                bias_smooth = np.zeros_like(bias, dtype=np.float64)
+                bias_smooth[0:h//2, :] = gaussian_filter(bias[0:h//2, :],
+                                            sigma = smooth_sigma,
+                                            mode  = smooth_mode)
+                bias_smooth[h//2:h, :] = gaussian_filter(bias[h//2:h, :],
+                                            sigma = smooth_sigma,
+                                            mode  = smooth_mode)
+
+                # write information to FITS header
+                bias_card_lst.append((prefix + 'SMOOTH CORRECTED',  True))
+                bias_card_lst.append((prefix + 'SMOOTH METHOD', 'GAUSSIAN'))
+                bias_card_lst.append((prefix + 'SMOOTH SIGMA',  smooth_sigma))
+                bias_card_lst.append((prefix + 'SMOOTH MODE',   smooth_mode))
+            else:
+                print('Unknown smooth method: ', smooth_method)
+                pass
+
+            bias = bias_smooth
+        else:
+            # bias not smoothed
+            bias_card_lst.append((prefix + 'SMOOTH CORRECTED', False))
+
+        # create new FITS Header for bias
+        head = fits.Header()
+        for card in bias_card_lst:
+            head.append(card)
+        fits.writeto(bias_file, bias, header=head, overwrite=True)
+
+        message = 'Bias image written to "{}"'.format(bias_file)
+        logger.info(message)
+        print(message)
+
+    return bias, bias_card_lst
+
+
+def select_calib_from_database(database_path, dateobs):
+    """Select wavelength calibration file in database.
+
+    Args:
+        path (str): Path to search for the calibration files.
+        dateobs (str): .
+    Returns:
+        tuple: A tuple containing:
+
+            * **spec** (:class:`numpy.dtype`): An array of previous calibrated
+                spectra.
+            * **calib** (dict): Previous calibration results.
+    """
+    
+    indexfile = os.path.join(database_path, 'index.dat')
+    calibtable = read_obslog(indexfile)
+
+    input_date = dateutil.parser.parse(dateobs)
+    if input_date > datetime.datetime(2019, 1, 1):
+        mask = calibtable['obsdate'] > datetime.datetime(2019, 1, 1)
+    else:                         
+        mask = calibtable['obsdate'] < datetime.datetime(2019, 1, 1)
+    
+    fileid = calibtable[mask][-1]['fileid']
+
+    filename = 'wlcalib.{}.fits'.format(fileid)
+    filepath = os.path.join(database_path, filename)
+         
+    # load spec, calib, and aperset from selected FITS file
+    f = fits.open(filepath)
+    head = f[0].header
+    spec = f[1].data
+    f.close()
+
+    calib = get_calib_from_header(head)
+
+    return spec, calib
+
+
+def smooth_aperpar_A(newx_lst, ypara, fitmask, group_lst, w):
+    """Smooth *A* of the four 2D profile parameters (*A*, *k*, *c*, *bkg*) of
+    the fiber flat-fielding.
+
+    Args:
+        newx_lst (:class:`numpy.ndarray`): Sampling pixels of the 2D profile.
+        ypara (:class:`numpy.ndarray`): Array of *A* at the sampling pixels.
+        fitmask (:class:`numpy.ndarray`): Mask array of **ypara**.
+        group_lst (list): Groups of (*x*:sub:`1`, *x*:sub:`2`, ... *x*:sub:`N`)
+            in each segment, where *x*:sub:`i` are indices in **newx_lst**.
+        w (int): Length of flat.
+
+    Returns:
+        tuple: A tuple containing:
+
+            * **aperpar** (:class:`numpy.ndarray`) – Reconstructed profile
+              paramters at all pixels.
+            * **xpiece_lst** (:class:`numpy.ndarray`) – Reconstructed profile
+              parameters at sampling pixels in **newx_lst** for plotting.
+            * **ypiece_res_lst** (:class:`numpy.ndarray`) – Residuals of profile
+              parameters at sampling pixels in **newx_lst** for plotting.
+            * **mask_rej_lst** (:class:`numpy.ndarray`) – Mask of sampling pixels
+              in **newx_lst** participating in fitting or smoothing.
+
+    See Also:
+
+        * :func:`gamse.echelle.flat.get_fiber_flat`
+        * :func:`smooth_aperpar_k`
+        * :func:`smooth_aperpar_c`
+        * :func:`smooth_aperpar_bkg`
+    
+    """
+
+    has_fringe_lst = []
+    aperpar = np.array([np.nan]*w)
+    xpiece_lst     = np.array([np.nan]*newx_lst.size)
+    ypiece_res_lst = np.array([np.nan]*newx_lst.size)
+    mask_rej_lst   = np.array([np.nan]*newx_lst.size)
+    allx = np.arange(w)
+    # the dtype of xpiece_lst and ypiece_lst is np.float64
+
+    # first try, scan every segment. find fringe by checking the local maximum
+    # points after smoothing. Meanwhile, save the smoothing results in case the
+    # smoothing will be used afterwards.
+    for group in group_lst:
+        i1, i2 = group[0], group[-1]
+        p1, p2 = newx_lst[i1], newx_lst[i2]
+        m = fitmask[group]
+        xpiece = newx_lst[group]
+        ypiece = ypara[group]
+        # now fill the NaN values in ypiece
+        if (~m).sum() > 0:
+            f = InterpolatedUnivariateSpline(xpiece[m], ypiece[m], k=3)
+            ypiece = f(xpiece)
+        # now xpiece and ypiece are ready
+
+        _m = np.ones_like(ypiece, dtype=np.bool)
+        for ite in range(3):
+            f = InterpolatedUnivariateSpline(xpiece[_m], ypiece[_m], k=3)
+            ypiece2 = f(xpiece)
+            win_len = (11, 21)[ypiece2.size>23]
+            ysmooth = savgol_filter(ypiece2, window_length=win_len, polyorder=3)
+            res = ypiece - ysmooth
+            std = res.std()
+            _new_m = np.abs(res) < 3*std
+
+            # prevent extrapolation at the boundaries
+            if _new_m.size > 3:
+                _new_m[0:3] = True
+                _new_m[-3:] = True
+            _new_m = _m*_new_m
+
+            if _new_m.sum() == _m.sum():
+                break
+            _m = _new_m
+        # now xpiece, ypiece, ypiece2, ysmooth, res, and _m have the same
+        # lengths and meanings on their positions of elements
+
+        f = InterpolatedUnivariateSpline(xpiece, ysmooth, k=3)
+        _x = np.arange(p1, p2+1)
+
+        aperpar[_x] = f(_x)
+        xpiece_lst[group] = xpiece
+        ypiece_res_lst[group] = res
+        mask_rej_lst[group] = ~_m
+
+        # find out if this order is affected by fringes, by checking the
+        # distribution of local maximum points
+        imax, ymax = get_local_minima(-ysmooth, window=5)
+        if len(imax) > 0:
+            x = xpiece[imax]
+        else:
+            x = []
+        # determine how many pixels in each bin.
+        # if w=4000, then 500 pix. if w=2000, then 250 pix.
+        npixbin = w//8
+        bins = np.linspace(p1, p2, int(p2-p1)/npixbin+2)
+        hist, _ = np.histogram(x, bins)
+
+        n_nonzerobins = np.nonzero(hist)[0].size
+        n_zerobins = hist.size - n_nonzerobins
+
+        if p2-p1<w/8 or n_zerobins<=1 or \
+            n_zerobins<n_nonzerobins or n_nonzerobins>=3:
+            # there's fringe
+            has_fringe = True
+        else:
+            # no fringe
+            has_fringe = False
+        has_fringe_lst.append(has_fringe)
+
+    # use global polynomial fitting if this order is affected by fringe and the
+    # following conditions are satisified
+    if len(group_lst) > 1 and newx_lst[group_lst[0][0]] < w/2 and \
+        newx_lst[group_lst[-1][-1]] > w/2 and \
+        has_fringe_lst.count(True) == len(has_fringe_lst):
+        # fit polynomial over the whole order
+
+        # prepare xpiece and y piece
+        xpiece = np.concatenate([newx_lst[group] for group in group_lst])
+        ypiece = np.concatenate([ypara[group] for group in group_lst])
+
+        # fit with poly
+        # determine the polynomial degree
+        xspan = xpiece[-1] - xpiece[0]
+        deg = (((1, 2)[xspan>w/8], 3)[xspan>w/4], 4)[xspan>w/2]
+        coeff, ypiece_fit, ypiece_res, _m, std = iterative_polyfit(
+            xpiece/w, ypiece, deg=deg, maxiter=10, lower_clip=3,
+            upper_clip=3)
+        aperpar = np.polyval(coeff, allx/w)
+        xpiece_lst = xpiece
+        ypiece_res_lst = ypiece_res
+        mask_rej_lst = ~_m
+    else:
+        # scan again
+        # fit polynomial for every segment
+        for group, has_fringe in zip(group_lst, has_fringe_lst):
+            xpiece = newx_lst[group]
+            ypiece = ypara[group]
+            xspan = xpiece[-1] - xpiece[0]
+            if has_fringe:
+                deg = (((1, 2)[xspan>w/8], 3)[xspan>w/4], 4)[xspan>w/2]
+            else:
+                deg = 7
+            coeff, ypiece_fit, ypiece_res, _m, std = iterative_polyfit(
+                xpiece/w, np.log(ypiece), deg=deg, maxiter=10,
+                lower_clip=3, upper_clip=3)
+            ypiece_fit = np.exp(ypiece_fit)
+            ypiece_res = ypiece - ypiece_fit
+
+            ii = np.arange(xpiece[0], xpiece[-1]+1)
+            aperpar[ii] = np.exp(np.polyval(coeff, ii/w))
+            xpiece_lst[group]     = xpiece
+            ypiece_res_lst[group] = ypiece_res
+            mask_rej_lst[group]   = ~_m
+
+    return aperpar, xpiece_lst, ypiece_res_lst, mask_rej_lst
+
+def smooth_aperpar_k(newx_lst, ypara, fitmask, group_lst, w):
+    """Smooth *k* of the four 2D profile parameters (*A*, *k*, *c*, *bkg*) of
+    the fiber flat-fielding.
+
+    Args:
+        newx_lst (:class:`numpy.ndarray`): Sampling pixels of the 2D profile.
+        ypara (:class:`numpy.ndarray`): Array of *A* at the sampling pixels.
+        fitmask (:class:`numpy.ndarray`): Mask array of **ypara**.
+        group_lst (list): Groups of (*x*:sub:`1`, *x*:sub:`2`, ... *x*:sub:`N`)
+            in each segment, where *x*:sub:`i` are indices in **newx_lst**.
+        w (int): Length of flat.
+
+    Returns:
+        tuple: A tuple containing:
+
+            * **aperpar** (:class:`numpy.ndarray`) – Reconstructed profile
+              paramters at all pixels.
+            * **xpiece_lst** (:class:`numpy.ndarray`) – Reconstructed profile
+              parameters at sampling pixels in **newx_lst** for plotting.
+            * **ypiece_res_lst** (:class:`numpy.ndarray`) – Residuals of profile
+              parameters at sampling pixels in **newx_lst** for plotting.
+            * **mask_rej_lst** (:class:`numpy.ndarray`) – Mask of sampling pixels
+              in **newx_lst** participating in fitting or smoothing.
+
+    See Also:
+
+        * :func:`gamse.echelle.flat.get_fiber_flat`
+        * :func:`smooth_aperpar_A`
+        * :func:`smooth_aperpar_c`
+        * :func:`smooth_aperpar_bkg`
+    """
+
+    allx = np.arange(w)
+
+    if len(group_lst) > 1 and newx_lst[group_lst[0][0]] < w/2 and \
+        newx_lst[group_lst[-1][-1]] > w/2:
+        # fit polynomial over the whole order
+        xpiece = np.concatenate([newx_lst[group] for group in group_lst])
+        ypiece = np.concatenate([ypara[group] for group in group_lst])
+
+        xspan = xpiece[-1] - xpiece[0]
+        deg = (((1, 2)[xspan>w/8], 3)[xspan>w/4], 4)[xspan>w/2]
+
+        coeff, ypiece_fit, ypiece_res, _m, std = iterative_polyfit(
+            xpiece/w, ypiece, deg=deg, maxiter=10,
+            lower_clip=3, upper_clip=3)
+
+        aperpar = np.polyval(coeff, allx/w)
+        xpiece_lst     = xpiece
+        ypiece_res_lst = ypiece_res
+        mask_rej_lst   = ~_m
+    else:
+        # fit polynomial for every segment
+        aperpar = np.array([np.nan]*w)
+        xpiece_lst     = np.array([np.nan]*newx_lst.size)
+        ypiece_res_lst = np.array([np.nan]*newx_lst.size)
+        mask_rej_lst   = np.array([np.nan]*newx_lst.size)
+        for group in group_lst:
+            xpiece = newx_lst[group]
+            ypiece = ypara[group]
+            xspan = xpiece[-1] - xpiece[0]
+            deg = (((1, 2)[xspan>w/8], 3)[xspan>w/4], 4)[xspan>w/2]
+
+            coeff, ypiece_fit, ypiece_res, _m, std = iterative_polyfit(
+                xpiece/w, ypiece, deg=deg, maxiter=10, lower_clip=3,
+                upper_clip=3)
+
+            ii = np.arange(xpiece[0], xpiece[-1]+1)
+            aperpar[ii] = np.polyval(coeff, ii/w)
+            xpiece_lst[group]     = xpiece
+            ypiece_res_lst[group] = ypiece_res
+            mask_rej_lst[group]   = ~_m
+
+    return aperpar, xpiece_lst, ypiece_res_lst, mask_rej_lst
+
+def smooth_aperpar_c(newx_lst, ypara, fitmask, group_lst, w):
+    """Smooth *c* of the four 2D profile parameters (*A*, *k*, *c*, *bkg*) of
+    the fiber flat-fielding.
+
+    Args:
+        newx_lst (:class:`numpy.ndarray`): Sampling pixels of the 2D profile.
+        ypara (:class:`numpy.ndarray`): Array of *A* at the sampling pixels.
+        fitmask (:class:`numpy.ndarray`): Mask array of **ypara**.
+        group_lst (list): Groups of (*x*:sub:`1`, *x*:sub:`2`, ... *x*:sub:`N`)
+            in each segment, where *x*:sub:`i` are indices in **newx_lst**.
+        w (int): Length of flat.
+
+    Returns:
+        tuple: A tuple containing:
+
+            * **aperpar** (:class:`numpy.ndarray`) – Reconstructed profile
+              paramters at all pixels.
+            * **xpiece_lst** (:class:`numpy.ndarray`) – Reconstructed profile
+              parameters at sampling pixels in **newx_lst** for plotting.
+            * **ypiece_res_lst** (:class:`numpy.ndarray`) – Residuals of profile
+              parameters at sampling pixels in **newx_lst** for plotting.
+            * **mask_rej_lst** (:class:`numpy.ndarray`) – Mask of sampling pixels
+              in **newx_lst** participating in fitting or smoothing.
+
+    See Also:
+
+        * :func:`gamse.echelle.flat.get_fiber_flat`
+        * :func:`smooth_aperpar_A`
+        * :func:`smooth_aperpar_k`
+        * :func:`smooth_aperpar_bkg`
+    """
+    return smooth_aperpar_k(newx_lst, ypara, fitmask, group_lst, w)
+
+def smooth_aperpar_bkg(newx_lst, ypara, fitmask, group_lst, w):
+    """Smooth *bkg* of the four 2D profile parameters (*A*, *k*, *c*, *bkg*) of
+    the fiber flat-fielding.
+
+    Args:
+        newx_lst (:class:`numpy.ndarray`): Sampling pixels of the 2D profile.
+        ypara (:class:`numpy.ndarray`): Array of *A* at the sampling pixels.
+        fitmask (:class:`numpy.ndarray`): Mask array of **ypara**.
+        group_lst (list): Groups of (*x*:sub:`1`, *x*:sub:`2`, ... *x*:sub:`N`)
+            in each segment, where *x*:sub:`i` are indices in **newx_lst**.
+        w (int): Length of flat.
+
+    Returns:
+        tuple: A tuple containing:
+
+            * **aperpar** (:class:`numpy.ndarray`) – Reconstructed profile
+              paramters at all pixels.
+            * **xpiece_lst** (:class:`numpy.ndarray`) – Reconstructed profile
+              parameters at sampling pixels in **newx_lst** for plotting.
+            * **ypiece_res_lst** (:class:`numpy.ndarray`) – Residuals of profile
+              parameters at sampling pixels in **newx_lst** for plotting.
+            * **mask_rej_lst** (:class:`numpy.ndarray`) – Mask of sampling pixels
+              in **newx_lst** participating in fitting or smoothing.
+
+    See Also:
+
+        * :func:`gamse.echelle.flat.get_fiber_flat`
+        * :func:`smooth_aperpar_A`
+        * :func:`smooth_aperpar_k`
+        * :func:`smooth_aperpar_c`
+    """
+
+    allx = np.arange(w)
+
+    # fit for bkg
+    if len(group_lst) > 1 and newx_lst[group_lst[0][0]] < w/2 and \
+        newx_lst[group_lst[-1][-1]] > w/2:
+        # fit polynomial over the whole order
+        xpiece = np.concatenate([newx_lst[group] for group in group_lst])
+        ypiece = np.concatenate([ypara[group] for group in group_lst])
+
+        xspan = xpiece[-1] - xpiece[0]
+        deg = (((1, 2)[xspan>w/8], 3)[xspan>w/4], 4)[xspan>w/2]
+
+        coeff, ypiece_fit, ypiece_res, _m, std = iterative_polyfit(
+            xpiece/w, ypiece, deg=deg, maxiter=10,
+            lower_clip=3, upper_clip=3)
+
+        aperpar = np.polyval(coeff, allx/w)
+        xpiece_lst     = xpiece
+        ypiece_res_lst = ypiece_res
+        mask_rej_lst   = ~_m
+    else:
+        # fit polynomial for every segment
+        aperpar = np.array([np.nan]*w)
+        xpiece_lst     = np.array([np.nan]*newx_lst.size)
+        ypiece_res_lst = np.array([np.nan]*newx_lst.size)
+        mask_rej_lst   = np.array([np.nan]*newx_lst.size)
+        for group in group_lst:
+            xpiece = newx_lst[group]
+            ypiece = ypara[group]
+            xspan = xpiece[-1] - xpiece[0]
+            deg = (((1, 2)[xspan>w/8], 3)[xspan>w/4], 7)[xspan>w/2]
+
+            scale = ('linear','log')[(ypiece<=0).sum()==0]
+            if scale=='log':
+                ypiece = np.log(ypiece)
+
+            coeff, ypiece_fit, ypiece_res, _m, std = iterative_polyfit(
+                xpiece/w, ypiece, deg=deg, maxiter=10, lower_clip=3,
+                upper_clip=3)
+
+            if scale=='log':
+                ypiece = np.exp(ypiece)
+                ypiece_fit = np.exp(ypiece_fit)
+                ypiece_res = ypiece - ypiece_fit
+
+            ii = np.arange(xpiece[0], xpiece[-1]+1)
+            aperpar[ii] = np.polyval(coeff, ii/w)
+            if scale=='log':
+                aperpar[ii] = np.exp(aperpar[ii])
+            xpiece_lst[group]     = xpiece
+            ypiece_res_lst[group] = ypiece_res
+            mask_rej_lst[group]   = ~_m
+
+    return aperpar, xpiece_lst, ypiece_res_lst, mask_rej_lst
+
+class TraceFigure(TraceFigureCommon):
+    """Figure to plot the order tracing.
+    """
+    def __init__(self):
+        TraceFigureCommon.__init__(self, figsize=(20,10), dpi=150)
+        self.ax1 = self.add_axes([0.05,0.07,0.43,0.86])
+        self.ax2 = self.add_axes([0.52,0.50,0.43,0.40])
+        self.ax3 = self.add_axes([0.52,0.10,0.43,0.40])
+        self.ax4 = self.ax3.twinx()
+
+class Xinglong216HRS(Reduction):
+
+    def __init__(self):
+        super(Xinglong216HRS, self).__init__(instrument='Xinglong216HRS')
+
+    def config_ccd(self):
+        '''Set CCD images configurations.
+        '''
+        self.ccd_config
+
+    def overscan(self):
+        '''
+        Overscan correction for Xinglong 2.16m Telescope HRS.
+
+        .. csv-table:: Accepted options in config file
+           :header: Option, Type, Description
+           :widths: 20, 10, 50
+
+           **skip**,    *bool*, Skip this step if *yes* and **mode** = *'debug'*.
+           **suffix**,  *str*,  Suffix of the corrected files.
+           **plot**,    *bool*, Plot the overscan levels if *yes*.
+           **var_fig**, *str*,  Filename of the overscan variation figure.
+
+        '''
+
+
+        def fix_cr(a):
+            m = a.mean(dtype=np.float64)
+            s = a.std(dtype=np.float64)
+            mask = a > m + 3.*s
+            if mask.sum()>0:
+                x = np.arange(a.size)
+                f = InterpolatedUnivariateSpline(x[~mask],a[~mask],k=3)
+                return f(x)
+            else:
+                return a
+        
+        # find output suffix for fits
+        self.output_suffix = self.config.get('overscan', 'suffix')
+
+        if self.config.getboolean('overscan', 'skip'):
+            logger.info('Skip [overscan] according to the config file')
+            self.input_suffix = self.output_suffix
+            return True
+
+        # keywords for mask
+        saturation_adu = 65535
+
+        # path alias
+        rawdata = self.paths['rawdata']
+        midproc = self.paths['midproc']
+        report  = self.paths['report']
+
+        # loop over all files (bias, dark, ThAr, flat...)
+        # to correct for the overscan
+
+        # prepare the item list
+        item_lst = [item for item in self.log]
+
+        for i, item in enumerate(item_lst):
+            logger.info('Correct overscan for item %3d: "%s"'%(
+                         item.frameid, item.fileid))
+
+            # read FITS data
+            filename = '%s%s.fits'%(item.fileid, self.input_suffix)
+            filepath = os.path.join(rawdata, filename)
+            data, head = fits.getdata(filepath, header=True)
+
+            h, w = data.shape
+            x1, x2 = w-head['COVER'], w
+
+            # find the overscan level along the y-axis
+            ovr_lst1 = data[0:h//2,x1+2:x2].mean(dtype=np.float64, axis=1)
+            ovr_lst2 = data[h//2:h,x1+2:x2].mean(dtype=np.float64, axis=1)
+
+            ovr_lst1_fix = fix_cr(ovr_lst1)
+            ovr_lst2_fix = fix_cr(ovr_lst2)
+
+            # apply the sav-gol fitler to the mean of overscan
+            ovrsmooth1 = savgol_filter(ovr_lst1_fix, window_length=301, polyorder=3)
+            ovrsmooth2 = savgol_filter(ovr_lst2_fix, window_length=301, polyorder=3)
+
+            # plot the overscan regions
+            if i%5 == 0:
+                fig = plt.figure(figsize=(10,6), dpi=150)
+
+            ax1 = fig.add_axes([0.08, 0.83-(i%5)*0.185, 0.42, 0.15])
+            ax2 = fig.add_axes([0.55, 0.83-(i%5)*0.185, 0.42, 0.15])
+
+            ax1.plot([0,0],[ovr_lst1_fix.min(), ovr_lst1_fix.max()], 'w-', alpha=0)
+            _y1, _y2 = ax1.get_ylim()
+            ax1.plot(np.arange(0, h//2), ovr_lst1, 'r-', alpha=0.3)
+            ax1.set_ylim(_y1, _y2)
+
+            ax2.plot([0,0],[ovr_lst2_fix.min(), ovr_lst2_fix.max()], 'w-', alpha=0)
+            _y1, _y2 = ax2.get_ylim()
+            ax2.plot(np.arange(h//2, h), ovr_lst2, 'b-', alpha=0.3)
+            ax2.set_ylim(_y1, _y2)
+
+            ax1.plot(np.arange(0, h//2), ovrsmooth1, 'm', ls='-')
+            ax2.plot(np.arange(h//2, h), ovrsmooth2, 'c', ls='-')
+            ax1.set_ylabel('ADU')
+            ax2.set_ylabel('')
+            ax1.set_xlim(0, h//2-1)
+            ax2.set_xlim(h//2, h-1)
+            for ax in [ax1, ax2]:
+                _x1, _x2 = ax.get_xlim()
+                _y1, _y2 = ax.get_ylim()
+                _x = 0.95*_x1 + 0.05*_x2
+                _y = 0.20*_y1 + 0.80*_y2
+                ax.text(_x, _y, item.fileid, fontsize=9)
+                for tick in ax.xaxis.get_major_ticks():
+                    tick.label1.set_fontsize(9)
+                for tick in ax.yaxis.get_major_ticks():
+                    tick.label1.set_fontsize(9)
+                ax.xaxis.set_major_formatter(tck.FormatStrFormatter('%g'))
+                ax.yaxis.set_major_formatter(tck.FormatStrFormatter('%g'))
+                ax.xaxis.set_major_locator(tck.MultipleLocator(500))
+                ax.xaxis.set_minor_locator(tck.MultipleLocator(100))
+            if i%5==4 or i==len(item_lst)-1:
+                ax1.set_xlabel('Y (pixel)')
+                ax2.set_xlabel('Y (pixel)')
+                figname = 'overscan_%02d.png'%(i//5+1)
+                figfile = os.path.join(report, figname)
+                fig.savefig(figfile)
+                logger.info('Save image: %s'%figfile)
+                plt.close(fig)
+
+            # determine shape of output image (also the shape of science region)
+            y1 = head['CRVAL2']
+            y2 = y1 + head['NAXIS2'] - head['ROVER']
+            ymid = (y1 + y2)//2
+            x1 = head['CRVAL1']
+            x2 = x1 + head['NAXIS1'] - head['COVER']
+            newshape = (y2-y1, x2-x1)
+
+            # find the saturation mask
+            mask_sat = (data[y1:y2,x1:x2]>=saturation_adu)
+            # get bad pixel mask
+            bins = (head['RBIN'], head['CBIN'])
+            mask_bad = self._get_badpixel_mask(newshape, bins=bins)
+
+            mask = np.int16(mask_sat)*4 + np.int16(mask_bad)*2
+            # save the mask
+            mask_table = array_to_table(mask)
+            maskname = '%s%s.fits'%(item.fileid, self.mask_suffix)
+            maskpath = os.path.join(midproc, maskname)
+            fits.writeto(maskpath, mask_table, overwrite=True)
+
+            # subtract overscan
+            new_data = np.zeros(newshape, dtype=np.float64)
+            ovrdata1 = np.transpose(np.repeat([ovrsmooth1],x2-x1,axis=0))
+            ovrdata2 = np.transpose(np.repeat([ovrsmooth2],x2-x1,axis=0))
+            new_data[y1:ymid, x1:x2] = data[y1:ymid,x1:x2] - ovrdata1
+            new_data[ymid:y2, x1:x2] = data[ymid:y2,x1:x2] - ovrdata2
+
+            # fix bad pixels
+            new_data = fix_pixels(new_data, mask_bad, 'x', 'linear')
+
+            # update fits header
+            head['HIERARCH GAMSE OVERSCAN']        = True
+            head['HIERARCH GAMSE OVERSCAN METHOD'] = 'smooth'
+
+            # save data
+            outname = '%s%s.fits'%(item.fileid, self.output_suffix)
+            outpath = os.path.join(midproc, outname)
+            fits.writeto(outpath, new_data, head, overwrite=True)
+            print('Correct Overscan {} -> {}'.format(filename, outname))
+
+
+        logger.info('Overscan corrected. Change suffix: %s -> %s'%
+                    (self.input_suffix, self.output_suffix))
+        self.input_suffix = self.output_suffix
+
+    def _get_badpixel_mask(self, shape, bins):
+        '''Get bad-pixel mask.
+
+
+        Args:
+            shape (tuple): Shape of the science data region.
+            bins (tuple): Number of pixel bins of (ROW, COLUMN).
+        Returns:
+            :class:`numpy.array`: Binary mask indicating the bad pixels. The
+                shape of the mask is the same as the input shape.
+
+        The bad pixels are found when readout mode = Left Top & Bottom.
+
+        '''
+        mask = np.zeros(shape, dtype=np.bool)
+        if bins == (1, 1) and shape == (4136, 4096):
+            h, w = shape
+
+            mask[349:352, 627:630] = True
+            mask[349:h//2, 628]    = True
+
+            mask[1604:h//2, 2452] = True
+
+            mask[280:284,3701]   = True
+            mask[274:h//2, 3702] = True
+            mask[272:h//2, 3703] = True
+            mask[274:282, 3704]  = True
+
+            mask[1720:1722, 3532:3535] = True
+            mask[1720, 3535]           = True
+            mask[1722, 3532]           = True
+            mask[1720:h//2,3533]       = True
+
+            mask[347:349, 4082:4084] = True
+            mask[347:h//2,4083]      = True
+
+            mask[h//2:2631, 1909] = True
+        else:
+            print('No bad pixel information for this CCD size.')
+            raise ValueError
+        return mask
+
+
+
+    def bias(self):
+        '''Bias corrrection for Xinglong 2.16m Telescope HRS.
+
+        .. csv-table:: Accepted options in config file
+           :header: Option, Type, Description
+           :widths: 20, 10, 50
+
+           **skip**,        *bool*,  Skip this step if *yes* and **mode** = *'debug'*.
+           **suffix**,      *str*,   Suffix of the corrected files.
+           **cosmic_clip**, *float*, Upper clipping threshold to remove cosmic-rays.
+           **file**,        *str*,   Name of bias file.
+
+        '''
+        self.output_suffix = self.config.get('bias', 'suffix')
+
+        if self.config.getboolean('bias', 'skip'):
+            logger.info('Skip [bias] according to the config file')
+            self.input_suffix = self.output_suffix
+            return True
+
+        midproc = self.paths['midproc']
+
+        bias_id_lst = self.find_bias()
+
+        if len(bias_id_lst) == 0:
+            # no bias frame found. quit this method.
+            # update suffix
+            logger.info('No bias found.')
+            return True
+
+        infile_lst = [os.path.join(midproc,
+                        '%s%s.fits'%(item.fileid, self.input_suffix))
+                        for item in self.log if item.frameid in bias_id_lst]
+
+        # import and stack all bias files in a data cube
+        tmp = [fits.getdata(filename, header=True) for filename in infile_lst]
+        all_data, all_head = list(zip(*tmp))
+        all_data = np.array(all_data)
+
+        if self.config.has_option('bias', 'cosmic_clip'):
+            # use sigma-clipping method to mask cosmic rays
+            cosmic_clip = self.config.getfloat('bias', 'cosmic_clip')
+
+            all_mask = np.ones_like(all_data, dtype=np.bool)
+
+            mask = (all_data == all_data.max(axis=0))
+        
+            niter = 0
+            maxiter = 2
+            while niter <= maxiter:
+                niter += 1
+                mdata = np.ma.masked_array(all_data, mask=mask)
+                # calculate mean and standard deviation
+                mean = mdata.mean(axis=0, dtype=np.float64).data
+                std  = mdata.std(axis=0, dtype=np.float64, ddof=1).data
+                # initialize new mask
+                new_mask = np.ones_like(mask)>0
+                # masking all the upper outliers.
+                for i in np.arange(all_data.shape[0]):
+                    new_mask[i,:,:] = all_data[i,:,:] > mean + cosmic_clip*std
+                mask = new_mask
+
+            mdata = np.ma.masked_array(all_data, mask=mask)
+            bias = mdata.mean(axis=0, dtype=np.float64).data
+        else:
+            # no sigma clipping
+            bias = alldata.mean(axis=0, dtype=np.float64)
+
+        # create new FITS Header for bias
+        head = fits.Header()
+        head['HIERARCH GAMSE BIAS NFILE'] = len(bias_id_lst)
+
+        # get final bias filename from the config file
+        bias_file = self.config.get('bias', 'file')
+
+        if self.config.has_option('bias', 'smooth_method'):
+            # perform smoothing for bias
+            smooth_method = self.config.get('bias', 'smooth_method')
+            smooth_method = smooth_method.strip().lower()
+
+            logger.info('Smoothing bias: %s'%smooth_method)
+
+            if smooth_method in ['gauss','gaussian']:
+                # perform 2D gaussian smoothing
+
+                smooth_sigma = self.config.getint('bias', 'smooth_sigma')
+                smooth_mode  = self.config.get('bias', 'smooth_mode')
+
+                logger.info('Smoothing bias: sigma = %f'%smooth_sigma)
+                logger.info('Smoothing bias: mode = %s'%smooth_mode)
+
+                from scipy.ndimage.filters import gaussian_filter
+                h, w = bias.shape
+                bias_smooth = np.zeros((h, w), dtype=np.float64)
+                bias_smooth[0:h/2,:] = gaussian_filter(bias[0:h/2,:],
+                                                       smooth_sigma,
+                                                       mode=smooth_mode)
+                bias_smooth[h/2:h,:] = gaussian_filter(bias[h/2:h,:],
+                                                       smooth_sigma,
+                                                       mode=smooth_mode)
+
+                logger.info('Smoothing bias: Update bias FITS header')
+
+                head['HIERARCH GAMSE BIAS SMOOTH']        = True
+                head['HIERARCH GAMSE BIAS SMOOTH METHOD'] = 'GAUSSIAN'
+                head['HIERARCH GAMSE BIAS SMOOTH SIGMA']  = smooth_sigma
+                head['HIERARCH GAMSE BIAS SMOOTH MODE']   = smooth_mode
+
+            else:
+                pass
+
+            # bias_data is a proxy for bias to be corrected for each frame
+            bias_data = bias_smooth
+
+            # plot comparison between un-smoothed and smoothed data
+            self.plot_bias_smooth(bias, bias_smooth)
+
+        else:
+            # no smoothing
+            logger.info('No smoothing parameter for bias. Skip bias smoothing')
+            head['HIERARCH GAMSE BIAS SMOOTH'] = False
+            bias_data = bias
+
+        # save the bias to FITS
+        fits.writeto(bias_file, bias_data, head, overwrite=True)
+        
+        self.plot_bias_variation(all_data, all_head, time_key='DATE-STA')
+
+        # finally all files are corrected for the bias
+        for item in self.log:
+            if item.frameid not in bias_id_lst:
+                infile  = '%s%s.fits'%(item.fileid, self.input_suffix)
+                outfile = '%s%s.fits'%(item.fileid, self.output_suffix)
+                inpath  = os.path.join(midproc, infile)
+                outpath = os.path.join(midproc, outfile)
+                data, head = fits.getdata(inpath, header=True)
+                data_new = data - bias_data
+                # write information into FITS header
+                head['HIERARCH GAMSE BIAS'] = True
+                # save the bias corrected data
+                fits.writeto(outpath, data_new, head, overwrite=True)
+                info = ['Correct bias for item no. %d.'%item.frameid,
+                        'Save bias corrected file: "%s"'%outpath]
+                logger.info((os.linesep+'  ').join(info))
+                print('Correct bias: {} => {}'.format(infile, outfile))
+
+        # update suffix
+        logger.info('Bias corrected. Change suffix: %s -> %s'%
+                    (self.input_suffix, self.output_suffix))
+        self.input_suffix = self.output_suffix
+        return True
+
+all_columns = [
+        ('frameid', 'int',   '{:^7s}',  '{0[frameid]:7d}'),
+        ('fileid',  'str',   '{:^12s}', '{0[fileid]:12s}'),
+        ('imgtype', 'str',   '{:^7s}',  '{0[imgtype]:^7s}'),
+        ('object',  'str',   '{:^12s}', '{0[object]:12s}'),
+        ('i2cell',  'bool',  '{:^6s}',  '{0[i2cell]!s: <6}'),
+        ('exptime', 'float', '{:^7s}',  '{0[exptime]:7g}'),
+        ('obsdate', 'time',  '{:^23s}', '{0[obsdate]:}'),
+        ('nsat',    'int',   '{:^10s}', '{0[nsat]:10d}'),
+        ('q95',     'int',   '{:^10s}', '{0[q95]:10d}'),
+        ]
+
+def print_wrapper(string, item):
+    """A wrapper for log printing for Xinglong216HRS pipeline.
+
+    Args:
+        string (str): The output string for wrapping.
+        item (:class:`astropy.table.Row`): The log item.
+
+    Returns:
+        str: The color-coded string.
+
+    """
+    imgtype = item['imgtype']
+    obj     = item['object']
+
+    if len(obj)>=4 and obj[0:4].lower()=='bias':
+        # bias images, use dim (2)
+        return '\033[2m'+string.replace('\033[0m', '')+'\033[0m'
+
+    elif imgtype=='sci':
+        # sci images, use highlights (1)
+        return '\033[1m'+string.replace('\033[0m', '')+'\033[0m'
+
+    elif len(obj)>=4 and obj[0:4].lower()=='thar':
+        # arc lamp, use light yellow (93)
+        return '\033[93m'+string.replace('\033[0m', '')+'\033[0m'
+    else:
+        return string
+
+def make_config():
+    """Generate a config file for reducing the data taken with Xinglong 2.16m
+    HRS.
+
+
+    """
+
+    # find date of data obtained
+    current_pathname = os.path.basename(os.getcwd())
+    guess_date = extract_date(current_pathname)
+
+    while(True):
+        if guess_date is None:
+            prompt = 'YYYYMMDD'
+        else:
+            prompt = guess_date
+
+        string = input('Date of observation [{}]: '.format(prompt))
+        input_date = extract_date(string)
+        if input_date is None:
+            if guess_date is None:
+                continue
+            else:
+                input_date = guess_date
+                break
+        else:
+            break
+   
+    input_datetime = datetime.datetime.strptime(input_date, '%Y-%m-%d')
+
+    # determine the fiber mode
+    while(True):
+        string = input(
+            'The data was obatined with Single fiber or Double fibers? [s/d]:')
+        if string == 's':
+            fibermode = 'single'
+            break
+        elif string == 'd':
+            fibermode = 'double'
+            break
+        else:
+            print('Invalid input: {}'.format(string))
+            continue
+
+    # create config object
+    config = configparser.ConfigParser()
+
+    config.add_section('data')
+
+    # determine the time-dependent keywords
+    if input_datetime > datetime.datetime(2009, 1, 1):
+        # since 2019 there's another type of FITS header
+        statime_key = 'DATE-OBS'
+        exptime_key = 'EXPOSURE'
+    else:
+        statime_key = 'DATE-STA'
+        exptime_key = 'EXPTIME'
+
+    config.set('data', 'telescope',   'Xinglong216')
+    config.set('data', 'instrument',  'HRS')
+    config.set('data', 'rawdata',     'rawdata')
+    config.set('data', 'statime_key', statime_key)
+    config.set('data', 'exptime_key', exptime_key)
+    config.set('data', 'direction',   'xr-')
+    config.set('data', 'fibermode',   fibermode)
+    if fibermode == 'double':
+        config.set('data', 'fiberoffset', str(-12))
+
+    config.add_section('reduce')
+    config.set('reduce', 'midproc',     'midproc')
+    config.set('reduce', 'report',      'report')
+    config.set('reduce', 'onedspec',    'onedspec')
+    config.set('reduce', 'mode',        'normal')
+    config.set('reduce', 'oned_suffix', 'ods')
+    config.set('reduce', 'fig_format',  'png')
+    
+    config.add_section('reduce.bias')
+    config.set('reduce.bias', 'bias_file',     '${reduce:midproc}/bias.fits')
+    config.set('reduce.bias', 'cosmic_clip',   str(10))
+    config.set('reduce.bias', 'maxiter',       str(5))
+    config.set('reduce.bias', 'smooth',        'yes')
+    config.set('reduce.bias', 'smooth_method', 'gaussian')
+    config.set('reduce.bias', 'smooth_sigma',  str(3))
+    config.set('reduce.bias', 'smooth_mode',   'nearest')
+
+    config.add_section('reduce.trace')
+    config.set('reduce.trace', 'minimum',    str(8))
+    config.set('reduce.trace', 'scan_step',  str(100))
+    config.set('reduce.trace', 'separation', '500:21, 3500:52')
+    config.set('reduce.trace', 'filling',    str(0.3))
+    config.set('reduce.trace', 'align_deg',  str(2))
+    config.set('reduce.trace', 'display',    'no')
+    config.set('reduce.trace', 'degree',     str(3))
+
+    config.add_section('reduce.flat')
+    config.set('reduce.flat', 'slit_step',       str(256))
+    config.set('reduce.flat', 'q_threshold',     str(50))
+    config.set('reduce.flat', 'mosaic_maxcount', str(50000))
+
+    config.add_section('reduce.wlcalib')
+    config.set('reduce.wlcalib', 'search_database', 'yes')
+    config.set('reduce.wlcalib', 'database_path',
+                                    '/opt/gamse/Xinglong216.HRS/wlcalib')
+    config.set('reduce.wlcalib', 'linelist',        'thar.dat')
+    config.set('reduce.wlcalib', 'use_prev_fitpar', 'yes')
+    config.set('reduce.wlcalib', 'window_size',     str(13))
+    config.set('reduce.wlcalib', 'xorder',          str(3))
+    config.set('reduce.wlcalib', 'yorder',          str(3))
+    config.set('reduce.wlcalib', 'maxiter',         str(5))
+    config.set('reduce.wlcalib', 'clipping',        str(3))
+    config.set('reduce.wlcalib', 'q_threshold',     str(10))
+
+    config.add_section('reduce.background')
+    config.set('reduce.background', 'ncols',    str(9))
+    config.set('reduce.background', 'distance', str(7))
+    config.set('reduce.background', 'yorder',   str(7))
+
+    config.add_section('reduce.extract')
+    config.set('reduce.extract', 'upper_limit', str(7))
+    config.set('reduce.extract', 'lower_limit', str(7))
+
+    # write to config file
+    filename = 'Xinglong216HRS.{}.cfg'.format(input_date)
+    outfile = open(filename, 'w')
+    for section in config.sections():
+        maxkeylen = max([len(key) for key in config[section].keys()])
+        outfile.write('[{}]'.format(section)+os.linesep)
+        fmt = '{{:{}s}} = {{}}'.format(maxkeylen)
+        for key, value in config[section].items():
+            outfile.write(fmt.format(key, value)+os.linesep)
+        outfile.write(os.linesep)
+    outfile.close()
+
+    print('Config file written to {}'.format(filename))
+    
+
+def make_obslog(path):
+    """Scan the raw data, and generated a log file containing the detail
+    information for each frame.
+
+    An ascii file will be generated after running.
+    The name of the ascii file is `YYYY-MM-DD.obslog`, where `YYYY-MM-DD` is the
+    date of the *first* FITS image in the data folder.
+    If the file name already exists, `YYYY-MM-DD.1.obslog`,
+    `YYYY-MM-DD.2.obslog` ... will be used as substituions.
+
+    Args:
+        path (str): Path to the raw FITS files.
+
+    """
+    cal_objects = ['bias', 'flat', 'dark', 'i2', 'thar']
+    regular_names = ('Bias', 'Flat', 'ThAr', 'I2')
+
+    # if the obsinfo file exists, read and pack the information
+    addinfo_lst = {}
+    obsinfo_file = 'obsinfo.txt'
+    has_obsinfo = os.path.exists(obsinfo_file)
+    if has_obsinfo:
+        #io_registry.register_reader('obslog', Table, read_obslog)
+        #addinfo_table = Table.read(obsinfo_file, format='obslog')
+        addinfo_table = read_obslog(obsinfo_file)
+        addinfo_lst = {row['frameid']:row for row in addinfo_table}
+        # prepare the difference list between real observation time and FITS
+        # time
+        real_obsdate_lst = []
+        delta_t_lst = []
+
+    # scan the raw files
+    fname_lst = sorted(os.listdir(path))
+
+    # prepare logtable
+    logtable = Table(dtype=[
+        ('frameid', 'i2'),  ('fileid', 'S12'),  ('imgtype', 'S3'),
+        ('object',  'S12'), ('i2cell', 'bool'), ('exptime', 'f4'),
+        ('obsdate', Time),  ('nsat',   'i4'),   ('q95',     'i4'),
+        ])
+
+    # prepare infomation to print
+    pinfo = FormattedInfo(all_columns,
+            ['frameid', 'fileid', 'imgtype', 'object', 'i2cell', 'exptime',
+            'obsdate', 'nsat', 'q95'])
+
+    # print header of logtable
+    print(pinfo.get_separator())
+    print(pinfo.get_title())
+    #print(pinfo.get_dtype())
+    print(pinfo.get_separator())
+
+    prev_frameid = -1
+    # start scanning the raw files
+    for fname in fname_lst:
+        if fname[-5:] != '.fits':
+            continue
+        fileid  = fname[0:-5]
+        filename = os.path.join(path, fname)
+        data, head = fits.getdata(filename, header=True)
+
+        # determine the science and overscan regions
+        naxis1 = head['NAXIS1']
+        naxis2 = head['NAXIS2']
+        x1 = head.get('CRVAL1', 0)
+        y1 = head.get('CRVAL2', 0)
+        # get science region along x axis
+        cover = head.get('COVER')
+        if cover is None:
+            if naxis1 >= 4096:
+                cover = naxis1 - 4096
+        # get science region along y axis
+        rover = head.get('ROVER')
+        if rover is None:
+            if naxis2 >= 4136:
+                rover = naxis2 - 4136
+
+        # get start and end indices of science region
+        y2 = y1 + naxis2 - rover
+        x2 = x1 + naxis1 - cover
+        data = data[y1:y2,x1:x2]
+
+        # find frame-id
+        frameid = int(fileid[8:])
+        if frameid <= prev_frameid:
+            print('Warning: frameid {} > prev_frameid {}'.format(
+                    frameid, prev_frameid))
+
+        # parse obsdate
+        if 'DATE-STA' in head:
+            obsdate = Time(head['DATE-STA'])
+        else:
+            obsdate = Time(head['DATE-OBS'])
+        if (frameid in addinfo_lst and 'obsdate' in addinfo_table.colnames
+            and addinfo_lst[frameid]['obsdate'] is not np.ma.masked):
+            real_obsdate = addinfo_lst[frameid]['obsdate'].datetime
+            file_obsdate = obsdate.datetime
+            delta_t = real_obsdate - file_obsdate
+            real_obsdate_lst.append(real_obsdate)
+            delta_t_lst.append(delta_t.total_seconds())
+
+        if 'EXPTIME' in head:
+            exptime = head['EXPTIME']
+        else:
+            exptime = head['EXPOSURE']
+
+        # parse object name
+        if 'OBJECT' in head:
+            objectname = head['OBJECT'].strip()
+        else:
+            objectname = ''
+        if (frameid in addinfo_lst and 'object' in addinfo_table.colnames
+            and addinfo_lst[frameid]['object'] is not np.ma.masked):
+            objectname = addinfo_lst[frameid]['object']
+        # change to regular name
+        for regname in regular_names:
+            if objectname.lower() == regname.lower():
+                objectname = regname
+                break
+
+        # parse I2 cell
+        i2cell = objectname.lower()=='i2'
+        if (frameid in addinfo_lst and 'i2cell' in addinfo_table.colnames
+            and addinfo_lst[frameid]['i2cell'] is not np.ma.masked):
+            i2cell = addinfo_lst[frameid]['i2cell']
+
+        imgtype = ('sci', 'cal')[objectname.lower().strip() in cal_objects]
+
+        # determine the total number of saturated pixels
+        saturation = (data>=65535).sum()
+
+        # find the 95% quantile
+        quantile95 = np.sort(data.flatten())[int(data.size*0.95)]
+
+        item = [frameid, fileid, imgtype, objectname, i2cell, exptime, obsdate,
+                saturation, quantile95]
+        logtable.add_row(item)
+        # get table Row object. (not elegant!)
+        item = logtable[-1]
+
+        # print log item with colors
+        string = pinfo.get_format(has_esc=False).format(item)
+        print(print_wrapper(string, item))
+
+        prev_frameid = frameid
+
+    print(pinfo.get_separator())
+    
+    # sort by obsdate
+    #logtable.sort('obsdate')
+
+    if has_obsinfo and len(real_obsdate_lst)>0:
+        # determine the time offset as median value
+        time_offset = np.median(np.array(delta_t_lst))
+        time_offset_dt = datetime.timedelta(seconds=time_offset)
+        # plot time offset
+        fig = plt.figure(figsize=(9, 6), dpi=100)
+        ax = fig.add_axes([0.12,0.16,0.83,0.77])
+        xdates = mdates.date2num(real_obsdate_lst)
+        ax.plot_date(xdates, delta_t_lst, 'o-', ms=6)
+        ax.axhline(y=time_offset, color='k', ls='--', alpha=0.6)
+        ax.set_xlabel('Log Time', fontsize=12)
+        ax.set_ylabel('Log Time - FTIS Time (sec)', fontsize=12)
+        x1, x2 = ax.get_xlim()
+        y1, y2 = ax.get_ylim()
+        ax.text(0.95*x1+0.05*x2, 0.1*y1+0.9*y2,
+                'Time offset = %d seconds'%time_offset, fontsize=14)
+        ax.set_xlim(x1, x2)
+        ax.set_ylim(y1, y2)
+        ax.grid(True, ls='-', color='w')
+        ax.set_facecolor('#eaeaf6')
+        ax.set_axisbelow(True)
+        ax.spines['bottom'].set_color('none')
+        ax.spines['left'].set_color('none')
+        ax.spines['top'].set_color('none')
+        ax.spines['right'].set_color('none')
+        for t in ax.xaxis.get_ticklines():
+            t.set_color('none')
+        for t in ax.yaxis.get_ticklines():
+            t.set_color('none')
+        ax.xaxis.set_major_formatter(mdates.DateFormatter('%m-%d %H:%M'))
+        #plt.setp(ax.get_xticklabels(), rotation=30)i
+        fig.autofmt_xdate()
+        for tick in ax.xaxis.get_major_ticks():
+            tick.label1.set_fontsize(10)
+        for tick in ax.yaxis.get_major_ticks():
+            tick.label1.set_fontsize(10)
+        fig.suptitle('Time Offsets Between Log and FITS', fontsize=15)
+        fig.savefig('obsdate_offset.png')
+        plt.close(fig)
+
+        # correct time offset
+        for row in logtable:
+            row['obsdate'] = row['obsdate'] + time_offset_dt
+
+    # determine filename of logtable.
+    # use the obsdate of the first frame
+    obsdate = logtable[0]['obsdate'].iso[0:10]
+    outname = '{}.obslog'.format(obsdate)
+    if os.path.exists(outname):
+        i = 0
+        while(True):
+            i += 1
+            outname = '{}.{}.obslog'.format(obsdate, i)
+            if not os.path.exists(outname):
+                outfilename = outname
+                break
+    else:
+        outfilename = outname
+
+    # save the logtable
+    outfile = open(outfilename, 'w')
+    outfile.write(pinfo.get_title()+os.linesep)
+    outfile.write(pinfo.get_dtype()+os.linesep)
+    outfile.write(pinfo.get_separator()+os.linesep)
+    for row in logtable:
+        outfile.write(pinfo.get_format().format(row)+os.linesep)
+    outfile.close()
